add multiple conversions to primitive number types to PdfArray - DEVSIX-1447
Autoported commit.
Original commit hash: [91590340f]

+++ itext/itext.forms/itext/forms/PdfAcroForm.cs
@@ -769,10 +769,10 @@ namespace iText.Forms {
                             canvas.OpenTag(tagRef);
                         }
                         PdfArray oldMatrix = xObject.GetPdfObject().GetAsArray(PdfName.Matrix);
-                        if (oldMatrix != null && iText.IO.Util.JavaUtil.ArraysEquals(oldMatrix.ToDoubleArray(), new double[] { 1, 
-                            0, 0, 1, 0, 0 })) {
+                        if (oldMatrix != null && iText.IO.Util.JavaUtil.ArraysEquals(oldMatrix.ToFloatArray(), new float[] { 1, 0, 
+                            0, 1, 0, 0 })) {
                             Rectangle boundingBox = xObject.GetBBox().ToRectangle();
-                            PdfArray newMatrixArray = new PdfArray(new double[] { box.GetWidth() / boundingBox.GetWidth(), 0, 0, box.GetHeight
+                            PdfArray newMatrixArray = new PdfArray(new float[] { box.GetWidth() / boundingBox.GetWidth(), 0, 0, box.GetHeight
                                 () / boundingBox.GetHeight(), 0, 0 });
                             xObject.Put(PdfName.Matrix, new PdfArray(newMatrixArray));
                         }
